Update README.md
The following data comes from the Consumer Complaint Database. Complaints may provide insight into the kinds of challenges consumers experience with financial products and services. The following charts take a deeper look into consumer complaint data in the state of California in the year 2020. It focuses primarily on the companies associated with the greatest number of complaints filed, as well as the kinds of issues filed. The data is meant to portray the between the top 3-10 of these trends.